requirements.txt
asgiref==3.8.1
attrs==23.2.0
automium==0.2.6
automium_web==0.1.1
azure-core==1.30.2
azure-storage-blob==12.21.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cheroot==10.0.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dataclasses==0.6
distlib==0.3.8
Django==4.2.14
django-cors-headers==4.4.0
django-cors-middleware==1.5.0
django-mongoengine==0.5.6
djangorestframework==3.15.2
dnspython==2.6.1
Faker==20.1.0
filelock==3.15.4
Flask==3.0.3
geckodriver-autoinstaller==0.1.0
h11==0.14.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
jaraco.functools==4.0.1
Jinja2==3.1.4
jsoncodable==0.1.7
jsonpickle==3.2.2
k-selenium-cookies==0.0.4
kproxy==0.0.1
legacy==0.1.7
MarkupSafe==2.1.5
mongoengine==0.28.2
more-itertools==10.3.0
noraise==0.0.16
outcome==1.3.0.post0
pexels==0.0.11
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
pymongo==3.12.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
selenium==4.23.0
selenium-browser==0.0.15
selenium-firefox==2.0.8
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.3.1
tldextract==5.1.2
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
web.py==0.62
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
xpath-utils==0.0.3